+++ ts/index.ts
@@ -7,7 +7,6 @@
 
 import {
   Iterable,
-  IterableConfig,
   IterableAction,
   IterableActionContext,
   IterableAttributionInfo,
@@ -29,6 +28,7 @@ import {
   IterableInAppDeleteSource,
 } from './IterableInAppClasses'
 
+import { IterableConfig } from './IterableConfig'
 import { IterableInAppMessage } from './IterableInAppMessage'
 import { IterableInAppManager } from './IterableInAppManager'
